backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
openpyxl
matplotlib
reportlab
Pillow
python-dotenv
supabase
